api/requirements.txt
fastapi>=0.80.0
uvicorn[standard]>=0.18.0
google-ortools==9.7.2996
pandas>=1.3.0
scikit-learn>=1.0
joblib>=1.0
pymongo[srv]>=4.0
python-dotenv>=0.20.0
httpx>=0.23.0

# Testing
pytest>=7.0.0
mongomock>=4.0.0
